Function rustdoc_to_markdown::process_docs

pub fn process_docs<S: Into<String>, L: Into<Vec<S>>>(
    lines: L, 
    indent_headings: bool
) -> Vec<String>

Process and concatenate the doc lines into a single String

The processing transforms doc tests into regular rust code blocks and optionally indent the markdown headings in order to leave the top heading to the crate name